dress_code-main123/scrap_files/requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
ultralytics>=8.0.0
pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
flask>=2.3.0
scikit-learn>=1.3.0
pyyaml>=6.0
joblib>=1.2.0
requests>=2.31.0
